+++ CHANGELOG.md
@@ -4,9 +4,16 @@
 # Assessment tool
 
 ---
+### v1.0.15
+> Updated: 18.05.2026
+1. Stop incorrect description inheritance from previous node
+
 ### v1.0.14
-> Updated: 12.05.2026
+> Updated: 18.05.2026
 1. Style competency statements as inset text
+2. Add framework variant attributes to assessment listing page
+3. Update ReadMe, etc.
+4. Package updates
 
 ---
 ### v1.0.13